Add Helmet.js for added HTTP header security
Added Helmet.js as a new dependency in package.json and package-lock.json, and imported and used it in index.js. Helmet.js is used to secure the Express app by setting various HTTP headers. This enhancement is part of our continuous effort to improve the security of our app.