Add ddos package for DOS attack prevention
Integrated 'ddos' module into our application to provide a mechanism for preventing Denial of Service (DoS) attacks. A new instance of Ddos (burst limit: 10, request limit: 15) has been constructed and added as a middleware function in our Express app to control the rate of incoming requests. This is a proactive measure taken for security enhancement.